Change the representation of arrow function bodies
The issue with the current parsing rules and representation is that
arrow function bodies are not statements, they're either function blocks
or expressions. Looking at the
specification (https://www.ecma-international.org/ecma-262/#prod-ExpressionBody)
we see that the expressions are `AssignmentExpression` in particular.

Making this change appears to fix the additional tests, but it breaks
minification.